Extract environment model
