+++ packages/insomnia-inso/src/commands/run-tests.js
@@ -3,10 +3,10 @@
 import { generate, runTestsCli } from 'insomnia-testing';
 import type { GlobalOptions } from '../util';
 import { loadDb } from '../db';
-import { loadEnvironment, promptEnvironment } from '../db/prompts';
 import type { UnitTest, UnitTestSuite } from '../db/types';
 import { noConsoleLog } from '../logger';
 import { loadTestSuites, promptTestSuites } from '../db/models/unit-test-suite';
+import { loadEnvironment, promptEnvironment } from '../db/models/environment';
 
 export const TestReporterEnum = {
   dot: 'dot',
